requirements.txt
Django==1.7.1
MySQL-python==1.2.5
argparse==1.2.1
cssmin==0.2.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.2
django-mockups==0.4.8
django-pipeline==1.3.26
django-registration-redux==1.1
djangorestframework==3.0.1
futures==2.2.0
ply==3.4
slimit==0.8.1
sqlparse==0.1.12
wsgiref==0.1.2
